# XPS_INTELLIGENCE_SYSTEM/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: XPS Intelligence API
  description: Backend API contract for XPS Intelligence autonomous AI platform
  version: 1.0.0
  contact:
    name: InfinityXOne Systems
    url: https://github.com/InfinityXOneSystems
servers:
  - url: http://localhost:3000/api
    description: Local development (Express gateway, default port)
  - url: http://localhost:3200/api
    description: Local development (Express gateway, legacy port)
  - url: http://localhost:8000
    description: Local development (FastAPI agent core)
  - url: https://api.xps-intelligence.com/api
    description: Production
...
paths:
  /leads:
    get:
      tags: [leads]
      summary: List all leads
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: minScore
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Array of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
    post:
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreate'
      responses:
        '201':
          description: Created lead
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
...
components:
  schemas:
    Lead:
      type: object
      required: [id, company, city, phone, email, rating, opportunityScore, status, createdAt]
      properties:
        id: { type: string }
        company: { type: string }
        city: { type: string }
        state: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        rating:
          type: string
          enum: ['A+', 'A', 'B+', 'B', 'C', 'D']
        opportunityScore: { type: number }
        lead_score: { type: number }
        assignedRep: { type: string }
        assignedInitials: { type: string }
        status:
          type: string
          enum: [new, contacted, qualified, proposal, signed, lost]
        priority:
          type: string
          enum: [green, yellow, red]
        isNew: { type: boolean }
        notes: { type: string }
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category: { type: string }
        address: { type: string }
        revenue: { type: number }
        lat: { type: number }
        lng: { type: number }
    LeadCreate:
      allOf:
        - $ref: '#/components/schemas/Lead'
        - type: object
          required: [company, city, phone, email, rating, opportunityScore, status]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []